Move longrunning prod deployment with tests to own file
